requirements.txt
cassandra-driver
pandas
requests
cffi